requirements-dev.txt
-r requirements.txt
pytest==8.3.0
pytest-django==4.9.0
pytest-asyncio==0.25.0
pytest-cov==6.0.0
ruff==0.8.0
